+++ src/checkout-buttons/strategies/apple-pay/index.ts
@@ -1,0 +1,2 @@
+export { default as ApplePayButtonStrategy } from './apple-pay-button-strategy';
+export { default as ApplePayButtonInitializeOptions } from './apple-pay-button-initialize-options';