+++ src/payment/create-payment-strategy-registry.ts
@@ -56,7 +56,7 @@ import { MonerisPaymentStrategy } from './strategies/moneris';
 import { NoPaymentDataRequiredPaymentStrategy } from './strategies/no-payment';
 import { OfflinePaymentStrategy } from './strategies/offline';
 import { OffsitePaymentStrategy } from './strategies/offsite';
-import { OpyPaymentStrategy } from './strategies/opy';
+import { OpyPaymentStrategy, OpyScriptLoader } from './strategies/opy';
 import { PaypalExpressPaymentStrategy, PaypalProPaymentStrategy, PaypalScriptLoader } from './strategies/paypal';
 import { createPaypalCommercePaymentProcessor,
     PaypalCommerceCreditCardPaymentStrategy,
@@ -561,7 +561,8 @@ export default function createPaymentStrategyRegistry(
             orderActionCreator,
             paymentMethodActionCreator,
             storefrontPaymentRequestSender,
-            paymentActionCreator
+            paymentActionCreator,
+            new OpyScriptLoader(scriptLoader)
         )
     );